refactor ,command covered with tests

+++ src/main/java/ua/alexander/sqlcmd/controller/command/Drop.java
@@ -1,6 +1,6 @@
 package ua.alexander.sqlcmd.controller.command;
 
-import ua.alexander.sqlcmd.controller.tools.CommandTools;
+import ua.alexander.sqlcmd.controller.tools.CommandTool;
 import ua.alexander.sqlcmd.module.DataBaseManager;
 import ua.alexander.sqlcmd.view.View;
 
@@ -8,10 +8,8 @@ public class Drop implements Command {
     private static final String COMMAND_SAMPLE = "drop:user";
     private DataBaseManager dbManager;
     private View view;
-    private CommandTools tool;
 
     public Drop(View view, DataBaseManager dbManager) {
-        tool = new CommandTools();
         this.view = view;
         this.dbManager = dbManager;
     }
@@ -23,18 +21,15 @@ public class Drop implements Command {
 
     @Override
     public void execute(String command) {
+        String[] data = command.split("[:]");
+        CommandTool.getCommandTool().validateCommandWithFixedSize(data, COMMAND_SAMPLE);
         view.type("Are you sure you want to delete the table? Type 'y' to confirm, 'n' to discard");
-        if (new CommandTools().verification())
-            executeDropping(command);
+        if (CommandTool.getCommandTool().verification()) {
+            String tableName = data[1];
+            dbManager.dropTable(tableName);
+            view.type(String.format("Table '%s' was deleted successfully!", tableName));
+        }
         else
             view.type("Table wasn't deleted.");
-    }
-
-    public void executeDropping(String command) {
-        String[] data = command.split("[:]");
-        tool.validateCommandWithFixedSize(data, COMMAND_SAMPLE);
-        String tableName = data[1];
-        dbManager.dropTable(tableName);
-        view.type(String.format("Table '%s' was deleted successfully!", tableName));
     }
 }
